+++ service/controller/app/resource/configmap/resource.go
@@ -15,6 +15,8 @@ import (
 const (
 	// Name is the identifier of the resource.
 	Name = "configmap"
+
+	cordonedStatus = "cordoned"
 )
 
 // Config represents the configuration used to create a new configmap resource.